Moved notifications. Moved FormField to components
