+++ src/ui/notifications/components/NotificationNotice.svelte
